requirements.txt
boto3
azure-storage-blob
paramiko
pyyaml
pydantic
pytest
